Refactor ScrolledPictures.js Component to reduce boilerplate
